requirements/dev.txt
# SHA1:171ecd91e8d5cec040a3287f5d4b90edc3869349
#
# This file is autogenerated by pip-compile-multi
# To update, run:
#
#    pip-compile-multi
#
-r base.txt
appdirs==1.4.3            # via black
attrs==19.3.0             # via black, flake8-bugbear, pytest
backcall==0.1.0           # via ipython
bandit==1.6.2             # via flake8-bandit
black==19.10b0            # via flake8-black
certifi==2019.11.28       # via requests
chardet==3.0.4            # via requests
click==7.0                # via black, coreapi-cli, pip-compile-multi, pip-tools
coreapi-cli==1.0.9
coreapi==2.3.3
coreschema==0.0.4         # via coreapi
coverage==5.0.3           # via pytest-cov
decorator==4.4.1          # via ipython, traitlets
django-debug-toolbar==2.2
django-extensions==2.2.8
entrypoints==0.3          # via flake8
factory-boy==2.12.0
faker==4.0.1              # via factory-boy
flake8-bandit==2.1.2
flake8-black==0.1.1
flake8-bugbear==20.1.4
flake8-polyfill==1.0.2    # via flake8-bandit, pep8-naming
flake8==3.7.9
gitdb==4.0.2              # via gitpython
gitpython==3.1.0          # via bandit
idna==2.9                 # via requests
importlib-metadata==1.5.0  # via pluggy, pytest
invoke==1.4.1
ipdb==0.13.1
ipython-genutils==0.2.0   # via traitlets
ipython==7.12.0           # via ipdb
isort==4.3.21
itypes==1.1.0             # via coreapi
jedi==0.16.0              # via ipython
jinja2==2.11.1            # via coreschema
markupsafe==1.1.1         # via jinja2
mccabe==0.6.1             # via flake8
more-itertools==8.2.0     # via pytest
packaging==20.1           # via pytest
parso==0.6.2              # via jedi
pathspec==0.7.0           # via black
pbr==5.4.4                # via stevedore
pep8-naming==0.9.1
pexpect==4.8.0            # via ipython
pickleshare==0.7.5        # via ipython
pip-compile-multi==1.5.8
pip-tools==4.5.1          # via pip-compile-multi
pluggy==0.13.1            # via pytest
prompt-toolkit==3.0.3     # via ipython
ptyprocess==0.6.0         # via pexpect
py==1.8.1                 # via pytest
pycodestyle==2.5.0        # via flake8, flake8-bandit
pyflakes==2.1.1           # via flake8
pygments==2.5.2           # via ipython
pyparsing==2.4.6          # via packaging
pytest-cov==2.8.1
pytest-django==3.8.0
pytest==5.3.5             # via pytest-cov, pytest-django
python-dateutil==2.8.1    # via faker
pyyaml==5.3
regex==2020.2.20          # via black
requests==2.23.0          # via coreapi
six==1.14.0               # via bandit, django-extensions, packaging, pip-tools, python-dateutil, stevedore, traitlets
smmap==3.0.1              # via gitdb
stevedore==1.32.0         # via bandit
toml==0.10.0              # via black
toposort==1.5             # via pip-compile-multi
traitlets==4.3.3          # via ipython
typed-ast==1.4.1          # via black
uritemplate==3.0.1        # via coreapi
urllib3==1.25.8           # via requests
wcwidth==0.1.8            # via prompt-toolkit, pytest
zipp==3.0.0               # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
